fixed an error
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/graph@3960 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/class.R
@@ -42,7 +42,7 @@
                           nodeType="character",
                           toEdges="list",
                           fromEdges="list",
-                          edgeOrder="integer",
+                          edgeOrder="numeric",
                           label="character" ),
            prototype = list(nodeID=newID(), nodetype="unknown",
              edgeOrder=0, label=""),
